+++ src/elements/three-scene.ts
@@ -47,7 +47,7 @@ export class ThreeScene extends ThreeElement.for(Scene) {
 
     /* Set up rendering */
     this.render = this.render.bind(this)
-    this.game.addEventListener("rendertick", this.render)
+    this.game.emitter.on("rendertick", this.render)
 
     /* Start processing events */
     this.pointer!.start()
@@ -62,7 +62,7 @@ export class ThreeScene extends ThreeElement.for(Scene) {
 
   disconnectedCallback() {
     /* Unregister event handlers */
-    this.game.removeEventListener("rendertick", this.render)
+    this.game.emitter.off("rendertick", this.render)
     window.removeEventListener("resize", this.handleWindowResize)
 
     super.disconnectedCallback()