+++ web-test-runner.config.mjs
@@ -1,8 +1,17 @@
+import { importMapsPlugin } from "@web/dev-server-import-maps"
 import { esbuildPlugin } from "@web/dev-server-esbuild"
 
 export default {
   files: ["test"],
   nodeResolve: true,
-  esbuildTarget: "auto",
-  plugins: [esbuildPlugin({ ts: true })]
+  debug: false,
+  watch: false,
+  open: false,
+  plugins: [
+    importMapsPlugin(),
+    esbuildPlugin({
+      ts: true,
+      target: "auto"
+    })
+  ]
 }